create fab for uploading image
