+++ client/src/pages/Profile.js
@@ -1,7 +1,6 @@
 import React, { useState } from 'react'
 import { Tab } from '@material-ui/core'
 import { TabContext, TabList, TabPanel } from '@material-ui/lab'
-import Layout from '../components/Layout'
 import Feed from '../components/Feed'
 import { user } from '../constant'
 import classes from '../styles/profile.module.css'
@@ -13,7 +12,7 @@ const Profile = () => {
             username, posts, followers, following, short,
         } = user
         return (
-            <Layout>
+            <>
                 <div className={classes.profile}>
                     <img
                         src="https://images.unsplash.com/photo-1519052537078-e6302a4968d4?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80"
@@ -48,7 +47,7 @@ const Profile = () => {
                         Summary
                     </TabPanel>
                 </TabContext>
-            </Layout>
+            </>
         )
     }
     return <ProfilePage />
